Completed Board Review. Minor Changes to Motor Bus Layout
Deleted Motor 7 from Board. Renamed Motors 1-6 for other uses and easier reading. Auxillary is now on the right hand side of the board, As well as non motor busses. Front Middle, Back and Extra are all on one side to fit with Motor Controller Specs. Added Extra Silkscreen for a more beautiful look.